style: rename command_line files to cmdline
